made some auto progress, odometry code broken completely

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -50,7 +50,7 @@ public abstract class Robot extends LinearOpMode {
     double rightBackPower;
     boolean DriverOrientationDriveMode = true;
     boolean Driver1Leftbumper;
-    float startingangle;
+    double startingangle;
 
 
     public void straferAlgorithm(){
